Pull juniper/web-controller to absingla/web-controller

+++ webroot/config/physicaldevices/physicalinterfaces/ui/js/physicalinterfaces_config.js
@@ -20,6 +20,7 @@ function physicalInterfacesConfig() {
     var vmiDataSrc = [];
     var selAllRows;
     var ajaxTimeout = 300000;
+    var deleteMsg = 'Confirm to delete all Interface(s) for Physical Router (%s)';
     //Method Definations
     this.load = load;
     this.destroy = destroy;	
@@ -39,6 +40,7 @@ function physicalInterfacesConfig() {
     
     function initComponents() {
         //initializing the virtual routers Grid
+        var deleteInterfacesDropdownTemplate = contrail.getTemplate4Id('delete-interface-action-template');
         $("#gridPhysicalInterfaces").contrailGrid({
             header : {
                 title: {
@@ -47,7 +49,7 @@ function physicalInterfacesConfig() {
                     //icon : 'icon-list',
                     //iconCssClass : 'blue'                
                 },
-                customControls: ['<a id="btnDeletePhysicalInterface" class="disabled-link" title="Delete Interface(s)"><i class="icon-trash"></i></a>',
+                customControls: [deleteInterfacesDropdownTemplate(),
                     '<a id="btnCreatePhysicalInterface" title="Create Interface"><i class="icon-plus"></i></a>',
                     'Router: <div id="ddPhysicalRouters"/>',]
             }, 
@@ -509,14 +511,34 @@ function physicalInterfacesConfig() {
         
         $('#btnDeletePhysicalInterface').click(function(){
              if(!$(this).hasClass('disabled-link')) {
+                 $('#confirmMainDelete').find("#txtConfirm").text('Confirm Interface(s) delete');
                  $('#confirmMainDelete').modal('show');
              }
         });
+        $('#btnDeleteAllInterfaces').click(function(){
+            var pRouter = $('#ddPhysicalRouters').data('contrailDropdown');
+            var pRouterName = pRouter != null ? pRouter.text() : '';
+            deleteMsg = deleteMsg.replace('%s', pRouterName)
+            $('#confirmMainDelete').find("#txtConfirm").text(deleteMsg);
+            $('#confirmMainDelete').modal('show');
+        });
         $('#btnCnfDelMainPopupOK').click(function(args){
-            var selected_rows = gridPhysicalInterfaces.getCheckedRows();
             $('#confirmMainDelete').modal("hide");
-            deletePhysicalInterface(selected_rows);
-        });  
+            if($('#confirmMainDelete').find("#txtConfirm").text() == deleteMsg) {
+                $.allajax.abort();
+                doAjaxCall('/api/tenants/config/delete-all-interfaces?prUUID=' + currentUUID, 'DELETE', null,
+                    'successHandlerForDeleteAll', 'failureHandlerForDeleteAll', null, null, ajaxTimeout);
+            } else {
+                var selected_rows = gridPhysicalInterfaces.getCheckedRows();
+                deletePhysicalInterface(selected_rows);
+            }
+        });
+        window.successHandlerForDeleteAll =  function(result) {
+            fetchPhysicalInterfaces();
+        }
+        window.failureHandlerForDeleteAll =  function(error) {
+            fetchPhysicalInterfaces();
+        }
         $('#txtPhysicalInterfaceName').bind('blur', function(e) {
               if($('#ddType').data('contrailDropdown').value() === "logical") {
                   var infName = $('#txtPhysicalInterfaceName').val().trim();
@@ -858,8 +880,8 @@ function physicalInterfacesConfig() {
             }
             for(var j = 0; j < selectedServerDetails.length ; j++){
                 var vmiFqname = 'none';
-                vmiFqname = selectedServerDetails[j]['vmi_fq_name'];
-                vmiRefs.push({"to" : [vmiFqname[0], vmiFqname[1], vmiFqname[2]],"uuid": selectedServerDetails[j]['vmi_uuid']});
+                vmiFqname = selectedServerDetails[j]['fq_name'];
+                vmiRefs.push({"to" : [vmiFqname[0], vmiFqname[1], vmiFqname[2]],"uuid": selectedServerDetails[j]['uuid']});
             }
             //Logical interface directly under pRouter case
             if(pRouterDD.value() === parent.value()) {
@@ -1190,7 +1212,7 @@ function physicalInterfacesConfig() {
                 $("[id$=serverMac]").data('contrailCombobox').enable(false);
             }
         } else {
-            doAjaxCall('/api/tenants/config/virtual-network-internals/' + id,'GET', null, 'successHandlerForVNInternals', 'failureHandlerForVNInternals', null, null, ajaxTimeout);
+            doAjaxCall('/api/tenants/config/get-virtual-machine-details/?vn_uuid=' + id,'GET', null, 'successHandlerForVNInternals', 'failureHandlerForVNInternals', null, null, ajaxTimeout);
         }
     }
     
@@ -1198,16 +1220,18 @@ function physicalInterfacesConfig() {
         vmiDataSrc = [];
         if(result != null && result.length > 0) {
             for(var i = 0; i < result.length; i++) {
-                var vmi = result[i];
-                var txt = vmi.mac[0]; //+ ' (' + vmi.ip[0] + ')';
-                var fixedIp = vmi.ip != null && vmi.ip.length > 0 ? vmi.ip[0] : '';
-                var txtVMI = '';
-                if(fixedIp != '') {
-                    txtVMI = txt + ' (' + fixedIp + ')';
-                } else {
-                    txtVMI = txt; 
+                if(result[i]['virtual-machine-interface'] != null) {
+                    var vmi = result[i]['virtual-machine-interface'];
+                    var txt = getMacAddress(vmi.virtual_machine_interface_mac_addresses);
+                    var fixedIp = vmi.instance_ip_address != null && vmi.instance_ip_address.length > 0 ? vmi.instance_ip_address[0] : '';
+                    var txtVMI = '';
+                    if(fixedIp != '') {
+                        txtVMI = txt + ' (' + fixedIp + ')';
+                    } else {
+                        txtVMI = txt;
+                    }
+                    vmiDataSrc.push({text : txtVMI, value : JSON.stringify(vmi.fq_name), ip : fixedIp, data:JSON.stringify(vmi)});
                 }
-                vmiDataSrc.push({text : txtVMI, value : JSON.stringify(vmi.vmi_fq_name), ip : fixedIp, data:JSON.stringify(vmi)});
             }
             if ($("[id$=serverMac]").length > 0){
                 $("[id$=serverMac]").data('contrailCombobox').setData(vmiDataSrc);
@@ -1238,6 +1262,14 @@ function physicalInterfacesConfig() {
         }       
     }
     
+    function getMacAddress(obj) {
+        var mac = '';
+        if(obj != null && obj.mac_address != null && obj.mac_address.length > 0) {
+           mac = obj.mac_address[0];
+        }
+        return mac;
+    }
+
     window.failureHandlerForVNInternals = function(error){
         var r = arguments;
         fetchVirtualNetworks();
@@ -1640,10 +1672,15 @@ function physicalInterfacesConfig() {
         	configDetailTemplate.remove();
         	configDetailTemplate = $();
         }
+        var configDetailTemplate = $("#delete-interface-action-template");
+        if(isSet(configDetailTemplate)) {
+            configDetailTemplate.remove();
+            configDetailTemplate = $();
+        }
         var ddPhysicalRouters = $("#ddPhysicalRouters").data("contrailDropdown");
         if(isSet(ddPhysicalRouters)) {
-           ddPhysicalRouters.destroy(); 
-           ddPhysicalRouters = $();
+            ddPhysicalRouters.destroy();
+            ddPhysicalRouters = $();
         }        
     }
